requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.9
PyYAML==5.1
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.1
